Map water quality data for a selected year

Map water quality data for a selected year

show_sitemappep(dat, yrsel, param = c("chla", "sd"),
  bay_segment = c("Western", "Central", "Eastern"))

Arguments

dat
data frame of data returned by read_pepwq

yrsel
numeric for year to plot

param
chr string indicating which water quality value to plot, one of "chla" for chlorophyll and "sd" for secchi depth

bay_segment
chr string for the bay segment, one or all of "Western", "Central", or "Eastern"

Value

A leaflet object

Details

Year estimates for the selected parameter are based on median observations across months. The color ramp is reversed for Secchi observations.

See also

Examples

show_sitemappep(rawdat, yrsel = 2018)